requirements.txt
numpy==1.11.3
olefile==0.43
Pillow==4.0.0
pygame==1.9.2
PyOpenGL==3.1.0
PyOpenGL-accelerate==3.1.0
